chore: removed users cache
