--- .sqlx/query-c467aa589b04d75815556bf41a142728fd69d73846807cf230a5a95add35c73e.json
@@ -1,12 +1,0 @@
-{
-  "db_name": "SQLite",
-  "query": "\n                INSERT INTO users (id, email, auth_provider, auth_provider_id)\n                VALUES ($1, $2, $3, $4)\n            ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Right": 4
-    },
-    "nullable": []
-  },
-  "hash": "c467aa589b04d75815556bf41a142728fd69d73846807cf230a5a95add35c73e"
-}
