+++ src/domain/events/mod.rs
@@ -1,28 +1,25 @@
 use anyhow::{bail, Result};
 use serde::{Deserialize, Serialize};
 
-use super::{ports::Port, projects::Project, users::User};
+use super::{ports::Port, projects::Project};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(untagged)]
 #[allow(clippy::enum_variant_names)]
 pub enum Event {
     ProjectCreated(Project),
-    UserCreated(User),
     PortCreated(Port),
 }
 impl Event {
     pub fn key(&self) -> String {
         match self {
             Event::ProjectCreated(_) => "ProjectCreated".into(),
-            Event::UserCreated(_) => "UserCreated".into(),
             Event::PortCreated(_) => "PortCreated".into(),
         }
     }
     pub fn from_key(key: &str, payload: &[u8]) -> Result<Self> {
         let event = match key {
             "ProjectCreated" => Self::ProjectCreated(serde_json::from_slice(payload)?),
-            "UserCreated" => Self::UserCreated(serde_json::from_slice(payload)?),
             "PortCreated" => Self::PortCreated(serde_json::from_slice(payload)?),
             _ => bail!("Event key not implemented"),
         };
