+++ src/driven/cache/mod.rs
@@ -3,7 +3,6 @@ use std::path::Path;
 
 pub mod port;
 pub mod project;
-pub mod user;
 
 pub struct SqliteCache {
     db: sqlx::sqlite::SqlitePool,
