+++ src/driven/mod.rs
@@ -1,4 +1,4 @@
-pub mod auth0;
+pub mod auth;
 pub mod cache;
 pub mod k8s;
 pub mod kafka;
